🐳 feat(docker-compose): add configuration for nexterm (#572)
Adds a new Docker Compose configuration for the big-bear-nexterm
application. This includes the service definition, network, and volume
configuration. The big-bear-nexterm service uses the
germannewsmaker/nexterm:1.0.2-OPEN-PREVIEW Docker image and exposes
port 6989 on the host. The application data is stored in a named
volume called big_bear_nexterm_data.

📝 docs(readme): add YouTube video info for nexterm

Adds a new section in the README file to provide information about a
YouTube video related to the nexterm application.