🐻‍❄️ feat(umbrel): add docker-compose for Umbrel OS on Portainer
Adds a new `docker-compose.yml` file to the `how-to-install-umbrel-os-on-portainer` directory. This
file defines the service for running the Umbrel OS container on Portainer. The container is
configured to use the `dockurr/umbrel:1.3.0` image, with the necessary port and volume mappings.
Additionally, a README.md file is added to provide a link to the corresponding YouTube video.